move some subscription code into TwitchLibSubscriptionWatcher, make subscription monitoring optional
