+++ TPP.Core/Chat/TwitchChat.cs
@@ -9,7 +9,6 @@ using TPP.Common;
 using TPP.Core.Configuration;
 using TPP.Core.Moderation;
 using TPP.Core.Overlay;
-using TPP.Core.Overlay.Events;
 using TPP.Model;
 using TPP.Persistence;
 using TwitchLib.Client;
@@ -17,7 +16,6 @@ using TwitchLib.Client.Events;
 using TwitchLib.Client.Models;
 using TwitchLib.Communication.Clients;
 using TwitchLib.Communication.Events;
-using static TPP.Core.EventUtils;
 using OnConnectedEventArgs = TwitchLib.Client.Events.OnConnectedEventArgs;
 
 namespace TPP.Core.Chat
@@ -32,11 +30,9 @@ namespace TPP.Core.Chat
         private readonly string _channel;
         public readonly string ChannelId;
         private readonly IUserRepo _userRepo;
-        private readonly ISubscriptionProcessor _subscriptionProcessor;
         private readonly TwitchClient _twitchClient;
         public readonly TwitchApiProvider TwitchApiProvider;
-        private readonly TwitchLibSubscriptionWatcher _subscriptionWatcher;
-        private readonly OverlayConnection _overlayConnection;
+        private readonly TwitchLibSubscriptionWatcher? _subscriptionWatcher;
         private readonly TwitchChatSender _twitchChatSender;
         private readonly TwitchChatModeChanger _twitchChatModeChanger;
         private readonly TwitchChatExecutor _twitchChatExecutor;
@@ -57,8 +53,6 @@ namespace TPP.Core.Chat
             _channel = chatConfig.Channel;
             ChannelId = chatConfig.ChannelId;
             _userRepo = userRepo;
-            _subscriptionProcessor = subscriptionProcessor;
-            _overlayConnection = overlayConnection;
 
             TwitchApiProvider = new TwitchApiProvider(
                 loggerFactory,
@@ -86,16 +80,16 @@ namespace TPP.Core.Chat
             _twitchClient.OnConnectionError += OnConnectionError;
             _twitchClient.OnMessageReceived += MessageReceived;
             _twitchClient.OnWhisperReceived += WhisperReceived;
-            _subscriptionWatcher = new TwitchLibSubscriptionWatcher(
-                loggerFactory.CreateLogger<TwitchLibSubscriptionWatcher>(), _userRepo, _twitchClient, clock);
-            _subscriptionWatcher.Subscribed += OnSubscribed;
-            _subscriptionWatcher.SubscriptionGifted += OnSubscriptionGifted;
-
             _twitchChatSender = new TwitchChatSender(loggerFactory, TwitchApiProvider, chatConfig, useTwitchReplies);
             _twitchChatModeChanger = new TwitchChatModeChanger(
                 loggerFactory.CreateLogger<TwitchChatModeChanger>(), TwitchApiProvider, chatConfig);
             _twitchChatExecutor = new TwitchChatExecutor(loggerFactory.CreateLogger<TwitchChatExecutor>(),
                 TwitchApiProvider, chatConfig);
+
+            _subscriptionWatcher = chatConfig.MonitorSubscriptions
+                ? new TwitchLibSubscriptionWatcher(loggerFactory, _userRepo, _twitchClient, clock,
+                    subscriptionProcessor, _twitchChatSender, overlayConnection)
+                : null;
         }
 
         private Task Connected(object? sender, OnConnectedEventArgs e) => _twitchClient.JoinChannelAsync(_channel);
@@ -120,100 +114,6 @@ namespace TPP.Core.Chat
             await _twitchClient.DisconnectAsync();
         }
 
-        private static string BuildSubResponse(
-            ISubscriptionProcessor.SubResult subResult, User? gifter, bool isAnonymous)
-        {
-            static string BuildOkMessage(ISubscriptionProcessor.SubResult.Ok ok, User? gifter, bool isAnonymous)
-            {
-                string message = "";
-                if (ok.SubCountCorrected)
-                    message +=
-                        $"We detected that the amount of months subscribed ({ok.CumulativeMonths}) is lower than " +
-                        "our system expected. This happened due to erroneously detected subscriptions in the past. " +
-                        "Your account data has been adjusted accordingly, and you will receive your rewards normally. ";
-                if (ok.NewLoyaltyLeague > ok.OldLoyaltyLeague)
-                    message += $"You reached Loyalty League {ok.NewLoyaltyLeague}! ";
-                if (ok.DeltaTokens > 0)
-                    message += $"You gained T{ok.DeltaTokens} tokens! ";
-                if (gifter != null && isAnonymous)
-                    message += "An anonymous user gifted you a subscription!";
-                else if (gifter != null && !isAnonymous)
-                    message += $"{gifter.Name} gifted you a subscription!";
-                else if (ok.CumulativeMonths > 1)
-                    message += "Thank you for resubscribing!";
-                else
-                    message += "Thank you for subscribing!";
-                return message;
-            }
-
-            return subResult switch
-            {
-                ISubscriptionProcessor.SubResult.Ok ok => BuildOkMessage(ok, gifter, isAnonymous),
-                ISubscriptionProcessor.SubResult.SameMonth sameMonth =>
-                    $"We detected that you've already announced your resub for month {sameMonth.Month}, " +
-                    "and received the appropriate tokens. " +
-                    "If you believe this is in error, please contact a moderator so this can be corrected.",
-                _ => throw new ArgumentOutOfRangeException(nameof(subResult)),
-            };
-        }
-
-        private void OnSubscribed(object? sender, SubscriptionInfo e)
-        {
-            TaskToVoidSafely(_logger, async () =>
-            {
-                ISubscriptionProcessor.SubResult subResult = await _subscriptionProcessor.ProcessSubscription(e);
-                string response = BuildSubResponse(subResult, null, false);
-                await _twitchChatSender.SendWhisper(e.Subscriber, response);
-
-                await _overlayConnection.Send(new NewSubscriber
-                {
-                    User = e.Subscriber,
-                    Emotes = e.Emotes.Select(EmoteInfo.FromOccurence).ToImmutableList(),
-                    SubMessage = e.Message,
-                    ShareSub = true,
-                }, CancellationToken.None);
-            });
-        }
-
-        private void OnSubscriptionGifted(object? sender, SubscriptionGiftInfo e)
-        {
-            TaskToVoidSafely(_logger, async () =>
-            {
-                (ISubscriptionProcessor.SubResult subResult, ISubscriptionProcessor.SubGiftResult subGiftResult) =
-                    await _subscriptionProcessor.ProcessSubscriptionGift(e);
-
-                string subResponse = BuildSubResponse(subResult, e.Gifter, e.IsAnonymous);
-                await _twitchChatSender.SendWhisper(e.SubscriptionInfo.Subscriber, subResponse);
-
-                string subGiftResponse = subGiftResult switch
-                {
-                    ISubscriptionProcessor.SubGiftResult.LinkedAccount =>
-                        $"As you are linked to the account '{e.SubscriptionInfo.Subscriber.Name}' you have gifted to, " +
-                        "you have not received a token bonus. " +
-                        "The recipient account still gains the normal benefits however. Thanks for subscribing!",
-                    ISubscriptionProcessor.SubGiftResult.SameMonth { Month: var month } =>
-                        $"We detected that this gift sub may have been a repeated message for month {month}, " +
-                        "and you have already received the appropriate tokens. " +
-                        "If you believe this is in error, please contact a moderator so this can be corrected.",
-                    ISubscriptionProcessor.SubGiftResult.Ok { GifterTokens: var tokens } =>
-                        $"Thank you for your generosity! You received T{tokens} tokens for giving a gift " +
-                        "subscription. The recipient has been notified and awarded their token benefits.",
-                    _ => throw new ArgumentOutOfRangeException(nameof(subGiftResult))
-                };
-                if (!e.IsAnonymous)
-                    await _twitchChatSender.SendWhisper(e.Gifter,
-                        subGiftResponse); // don't respond to the "AnAnonymousGifter" user
-
-                await _overlayConnection.Send(new NewSubscriber
-                {
-                    User = e.SubscriptionInfo.Subscriber,
-                    Emotes = e.SubscriptionInfo.Emotes.Select(EmoteInfo.FromOccurence).ToImmutableList(),
-                    SubMessage = e.SubscriptionInfo.Message,
-                    ShareSub = false,
-                }, CancellationToken.None);
-            });
-        }
-
         public async Task Start(CancellationToken cancellationToken)
         {
             await _twitchClient.ConnectAsync();
@@ -226,9 +126,7 @@ namespace TPP.Core.Chat
             await _twitchClient.DisconnectAsync();
             await _twitchChatSender.DisposeAsync();
             _twitchClient.OnConnected -= Connected;
-            _subscriptionWatcher.Subscribed -= OnSubscribed;
-            _subscriptionWatcher.SubscriptionGifted -= OnSubscriptionGifted;
-            _subscriptionWatcher.Dispose();
+            _subscriptionWatcher?.Dispose();
             _twitchClient.OnMessageReceived -= MessageReceived;
             _twitchClient.OnWhisperReceived -= WhisperReceived;
             _logger.LogDebug("twitch chat is now fully shut down");
